Update to v7

+++ libreg/inc/libreg/lim.h
@@ -168,7 +168,7 @@ void regLimMeasInit(struct reg_lim_meas *lim_meas, float pos_lim, float neg_lim,
  *                                 value for reg_lim_meas::meas2_filter_factor.
  * @param[in]     iter_period      Iteration period, used to calculate the value for reg_lim_meas::meas2_filter_factor.
  */
-void regLimRmsInit(struct reg_lim_rms *lim_rms, float rms_warning, float rms_fault, float rms_tc, double iter_period);
+void regLimRmsInit(struct reg_lim_rms *lim_rms, float rms_warning, float rms_fault, float rms_tc, float iter_period);
 
 /*!
  * Initialise field/current reference limits. Field/current limits use the same
